feat(solana): adds shared folder
