+++ packages/solana/Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
-    "programs/*"
+    "programs/*",
+    "shared/*"
 ]
 resolver = "2"
