+++ packages/solana/programs/debridge-reporter/Cargo.toml
@@ -18,4 +18,5 @@ idl-build = ["anchor-lang/idl-build"]
 
 [dependencies]
 anchor-lang = "0.30.1"
-debridge-solana-sdk = { git = "ssh://git@github.com/debridge-finance/debridge-solana-sdk.git", version = "1.0.2" }
+debridge-solana-sdk = "1.0.1"
+slots = { path = "../../shared/slots" }
