+++ packages/solana/shared/slots/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "slots"
+version = "0.1.0"
+description = "Utilities functions to access Solana slots"
+edition = "2021"
+
+[features]
+default = []
+cpi = ["no-entrypoint"]
+no-entrypoint = []
+no-idl = []
+no-log-ix-name = []
+idl-build = ["anchor-lang/idl-build"]
+
+[dependencies]
+anchor-lang = "0.30.1"
